change Ration From int into double

+++ ParsKyanCrm.Domain/Entities/DataFormQuestionsOption.cs
@@ -10,6 +10,6 @@ namespace ParsKyanCrm.Domain.Entities
         public int Id { get; set; }
         public string Text { get; set; }
         public int DataFormQuestionsId { get; set; }
-        public int? Ratio { get; set; }
+        public double? Ratio { get; set; }
     }
 }
